# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/alpha/scenarios/{scenarioId}": {
      "get": {
        "operationId": "AlphaController_getScenario",
        "parameters": [
          {
            "name": "scenarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meta",
                    "data"
                  ],
                  "properties": {
                    "meta": {
                      "type": "object",
                      "description": "Present on every alpha response. Values are fixed; clients must reject anything else.",
                      "required": [
                        "releaseChannel",
                        "liveData",
                        "liveCapitalAuthority",
                        "evidenceStatus",
                        "authority"
                      ],
                      "properties": {
                        "releaseChannel": {
                          "type": "string",
                          "enum": [
                            "private-validation-alpha"
                          ]
                        },
                        "liveData": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "liveCapitalAuthority": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "evidenceStatus": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ONLY"
                          ]
                        },
                        "authority": {
                          "type": "string",
                          "enum": [
                            "NO_LIVE_CAPITAL"
                          ]
                        }
                      }
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "version",
                        "caseId",
                        "title",
                        "description",
                        "bundle",
                        "bundleKey",
                        "frameCount",
                        "evidence",
                        "frames"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "version": {
                          "type": "string"
                        },
                        "caseId": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "bundle": {
                          "type": "object",
                          "description": "`AlphaTradeBundle` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                          "required": [
                            "identity",
                            "strategy",
                            "evidenceStatus",
                            "authoritySpecVersion"
                          ],
                          "properties": {
                            "identity": {
                              "type": "object",
                              "description": "`BundleIdentity` from @zugrio/domain: versioned MethodProfile / TradeBundle / RegimeModel / TimeframeMap and fixture scope.",
                              "required": [
                                "methodProfile",
                                "tradeBundle",
                                "regimeModel",
                                "timeframeMap",
                                "scope"
                              ],
                              "additionalProperties": true
                            },
                            "evidenceStatus": {
                              "type": "string",
                              "enum": [
                                "VALIDATION_ONLY"
                              ]
                            }
                          },
                          "additionalProperties": true
                        },
                        "bundleKey": {
                          "type": "string",
                          "description": "@zugrio/domain bundleIdentityKey(bundle.identity)"
                        },
                        "frameCount": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "evidence": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "`EvidenceEvent` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                            "required": [
                              "id",
                              "kind",
                              "knownAt",
                              "value",
                              "source"
                            ],
                            "properties": {
                              "knownAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "additionalProperties": true
                          }
                        },
                        "frames": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "`ReplayFrame` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                            "required": [
                              "evaluatedAt"
                            ],
                            "properties": {},
                            "additionalProperties": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {},
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get a validation replay scenario with all frames",
        "tags": [
          "private-validation-alpha"
        ]
      }
    },
    "/v1/alpha/scenarios/{scenarioId}/frames/{frameIndex}": {
      "get": {
        "operationId": "AlphaController_getFrameDecisionCase",
        "parameters": [
          {
            "name": "scenarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "frameIndex",
            "required": true,
            "in": "path",
            "description": "Zero-based replay frame index",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meta",
                    "data"
                  ],
                  "properties": {
                    "meta": {
                      "type": "object",
                      "description": "Present on every alpha response. Values are fixed; clients must reject anything else.",
                      "required": [
                        "releaseChannel",
                        "liveData",
                        "liveCapitalAuthority",
                        "evidenceStatus",
                        "authority"
                      ],
                      "properties": {
                        "releaseChannel": {
                          "type": "string",
                          "enum": [
                            "private-validation-alpha"
                          ]
                        },
                        "liveData": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "liveCapitalAuthority": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "evidenceStatus": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ONLY"
                          ]
                        },
                        "authority": {
                          "type": "string",
                          "enum": [
                            "NO_LIVE_CAPITAL"
                          ]
                        }
                      }
                    },
                    "data": {
                      "type": "object",
                      "description": "`DecisionCase` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                      "required": [
                        "caseId",
                        "scenarioId",
                        "scenarioVersion",
                        "evaluationId",
                        "bundle",
                        "current",
                        "structuralState",
                        "outcome",
                        "stateReason",
                        "outcomeReason",
                        "annotations",
                        "authority",
                        "authorityClass",
                        "modelScored",
                        "history"
                      ],
                      "properties": {
                        "caseId": {
                          "type": "string"
                        },
                        "scenarioId": {
                          "type": "string"
                        },
                        "scenarioVersion": {
                          "type": "string"
                        },
                        "evaluationId": {
                          "type": "string"
                        },
                        "bundle": {
                          "type": "object",
                          "description": "`AlphaTradeBundle` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                          "required": [
                            "identity",
                            "strategy",
                            "evidenceStatus",
                            "authoritySpecVersion"
                          ],
                          "properties": {
                            "identity": {
                              "type": "object",
                              "description": "`BundleIdentity` from @zugrio/domain: versioned MethodProfile / TradeBundle / RegimeModel / TimeframeMap and fixture scope.",
                              "required": [
                                "methodProfile",
                                "tradeBundle",
                                "regimeModel",
                                "timeframeMap",
                                "scope"
                              ],
                              "additionalProperties": true
                            },
                            "evidenceStatus": {
                              "type": "string",
                              "enum": [
                                "VALIDATION_ONLY"
                              ]
                            }
                          },
                          "additionalProperties": true
                        },
                        "current": {
                          "type": "object",
                          "description": "`EvidenceSnapshot` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                          "required": [
                            "evaluatedAt",
                            "evidenceIds",
                            "evidenceRefs"
                          ],
                          "properties": {},
                          "additionalProperties": true
                        },
                        "structuralState": {
                          "type": "string",
                          "nullable": true,
                          "description": "decision-core StructuralState, or null"
                        },
                        "outcome": {
                          "type": "string",
                          "description": "decision-core DecisionOutcome"
                        },
                        "annotations": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "`ChartAnnotation` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                            "required": [
                              "id",
                              "kind",
                              "knownAt",
                              "evidenceId"
                            ],
                            "properties": {},
                            "additionalProperties": true
                          }
                        },
                        "authority": {
                          "type": "string",
                          "enum": [
                            "NO_LIVE_CAPITAL"
                          ]
                        },
                        "authorityClass": {
                          "type": "string",
                          "enum": [
                            "STRUCTURAL_ONLY"
                          ]
                        },
                        "modelScored": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "history": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "`DecisionEvent` from @zugrio/decision-core (packages/decision-core/src/types.ts).",
                            "required": [
                              "evaluationId",
                              "evaluatedAt",
                              "structuralState",
                              "outcome",
                              "stateReason",
                              "outcomeReason",
                              "changes",
                              "annotationIds"
                            ],
                            "properties": {
                              "evaluationId": {
                                "type": "string"
                              },
                              "evaluatedAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "structuralState": {
                                "type": "string",
                                "nullable": true
                              },
                              "outcome": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": true
                          }
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {},
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {},
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Deterministic decision-core DecisionCase for one replay frame (not persisted)",
        "tags": [
          "private-validation-alpha"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zugrio API — private validation alpha",
    "description": "Validation-only control plane. Every response carries NO_LIVE_CAPITAL / VALIDATION_ONLY metadata. No live market data, broker connection or trade execution exists behind this API.",
    "version": "0.1.0-alpha.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.zugrio.xyz"
    }
  ],
...
}